% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a1.R
\docType{data}
\name{a1}
\alias{a1}
\title{Zbior danych a1 zawiera wektor czasow dla 10 000 krotnego wywolania gry w SuperFarmer strategia SuperFarmerDA::strategia_maxrabbit.
Zbior danych stworzony w celu szybszego dzialania funkcji wizytowkaMEWA}
\description{
Zbior danych a1 zawiera wektor czasow dla 10 000 krotnego wywolania gry w SuperFarmer strategia SuperFarmerDA::strategia_maxrabbit.
Zbior danych stworzony w celu szybszego dzialania funkcji wizytowkaMEWA
}

