% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a33.R
\docType{data}
\name{a33}
\alias{a33}
\title{Zbior danych a1 zawiera wektor czasow dla 10 000 krotnego wywolania gry w SuperFarmer strategia SuperFarmerRCNK::strategia_yolo.
Zbior danych stworzony w celu szybszego dzialania funkcji wizytowkaMEWA}
\description{
Zbior danych a1 zawiera wektor czasow dla 10 000 krotnego wywolania gry w SuperFarmer strategia SuperFarmerRCNK::strategia_yolo.
Zbior danych stworzony w celu szybszego dzialania funkcji wizytowkaMEWA
}

