% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Rozklad czasu gry w SuperFarmera zadana strategia}
\usage{
badaj_gre(strategia)
}
\arguments{
\item{strategia}{strategia, ktora chcemy zagrać. W pakiecie do wyboru: 'strategy1', 'strategy2', 'strategy3'}
}
\value{
funkcja zwraca podstawowe statyki czasu gry dana strategia po 10000 krotnych powtorzeniu fukcji gra()
}
\description{
Funkcja badaj_gre() oblicza podstawowe statystyki dotyczace czasu gry w SupreFarmera zadana strategia,
 po wykonianiu funkcji gra() 10000.
}

