% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre2.R
\name{badaj_gre2}
\alias{badaj_gre2}
\title{Rozklad czasu gry w SuperFarmera zadana strategia}
\usage{
badaj_gre2(strategia)
}
\arguments{
\item{strategia}{strategia, ktora chcemy zagrac. W pakiecie do wyboru: strategy1, strategy2, strategy3}
}
\value{
funkcja zwraca wektor, w którym każda współrzędna jest czasem gry zadana strategia
}
\description{
Funkcja badaj_gre2() powtarza gre dana strategia 10000 razy.
}

