% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceny_w_krolikach.R
\docType{data}
\name{ceny_w_krolikach}
\alias{ceny_w_krolikach}
\title{Zbior zawierajacy domyslne wartosci zwierzat wyrazone w krolikach, zapisane jako nazwany wektor.}
\description{
Zbior zawierajacy domyslne wartosci zwierzat wyrazone w krolikach, zapisane jako nazwany wektor.
}

