% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czy_wygrana.R
\name{czy_wygrana}
\alias{czy_wygrana}
\title{Warunek wygrania w gre SuperFarmer}
\usage{
czy_wygrana(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca wartość logiczna TRUE, gdy gracz wygral i FALSE w przeciwnym wypadku
}
\description{
Funkcja czy_wygrana() sprawdza, czy zostal spelniony warunek wygranej w grze SuperFarmer
}

