% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Gra w SuperFarmera zadana strategia}
\usage{
gra(strategia = SuperFarmerPKEK::strategy2,
  kostka1 = SuperFarmerPKEK::kostka1, kostka2 = SuperFarmerPKEK::kostka2,
  StadoGracza = SuperFarmerPKEK::StadoGracza)
}
\arguments{
\item{strategia}{strategia, ktora chcemy zagrac. W pakiecie do wyboru: strategy1, strategy2, strategy3}

\item{kostka1}{zwierzeta znajdujace sie na kostce nr 1}

\item{kostka2}{zwierzeta znajdujace sie na kostce nr 2}

\item{StadoGracza}{stan stada gracza, zapisany jako nazwany wektor}
}
\value{
funkcja zwraca czas gry dana strategia
}
\description{
Funkcja gra() oblicza czas gry w grę SupreFarmer zadana strategia
}

