% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3_kup_kilka_tanszych.R
\name{kup_kilka_tanszych}
\alias{kup_kilka_tanszych}
\title{Kupowanie kilku tanszych zwierzat za drozsze.}
\usage{
kup_kilka_tanszych(StadoGracza, zwierze, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{zwierze}{zwierze, ktore chcemy sprzedac}

\item{ceny_w_krolikach}{wektor opisujacy wartosć zwierzat wyrażona w krolikach}
}
\value{
funkcja zwraca StadoGracza kupieniu tanszych zwierzat
}
\description{
Kupowanie kilku tanszych zwierzat za drozsze.
}

