% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_stado.R
\docType{data}
\name{max_stado}
\alias{max_stado}
\title{Zbior zawierajacy maksymalna liczbe zwierzat dostepna w grze, zapisana jako nazwany wektor.}
\description{
Zbior zawierajacy maksymalna liczbe zwierzat dostepna w grze, zapisana jako nazwany wektor.
}

