% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porownaj_strategie.R
\name{porownaj_strategie}
\alias{porownaj_strategie}
\title{Rozklad czasu gry w SuperFarmera dla trzech zadanych strategii}
\usage{
porownaj_strategie(strategy1 = SuperFarmerPKEK::strategy1,
  strategy2 = SuperFarmerPKEK::strategy2,
  strategy3 = SuperFarmerPKEK::strategy3)
}
\arguments{
\item{strategy1}{strategia, ktorej rozklad chcemy zobaczyc na wykresie skrzypcowym}

\item{strategy2}{strategia, ktorej rozklad chcemy zobaczyc na wykresie skrzypcowym}

\item{strategy3}{strategia, ktorej rozklad chcemy zobaczyc na wykresie skrzypcowym}
}
\value{
funkcja zwraca wykres skrzypcowy z rozkladem czasu gry
}
\description{
Funkcja porownaj_strategie przedstawia na wykresie skrzypcowym rozklad czasu gry dla trzech dowolnych strategii 
(także z innych pakietów), przy powtorzeniu gry 10000 razy. 
Domyslnie rysowane są wykresy dla trzech strategii dostepnych w pakiecie
}

