% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/porownaj_strategie_w_pakiecie.R
\name{porownaj_strategie_w_pakiecie}
\alias{porownaj_strategie_w_pakiecie}
\title{Rozklad czasu gry w SuperFarmera strategii dostepnych w pakiecie}
\usage{
porownaj_strategie_w_pakiecie()
}
\value{
funkcja zwraca wykres skrzypcowy z rozkladem czasu gry
}
\description{
Funkcja porownaj_strategie przedstawia na wykresie skrzypcowym rozklad czasu gry strategii dostepnych
w pakiecie, przy powtorzeniu gry 10000 razy
}

