% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{reproduction}
\alias{reproduction}
\title{Mechanizm gry, dodawanie zwierzat po rzucie kostkami}
\usage{
reproduction(kostka1.wynik, kostka2.wynik, StadoGracza,
  max_stado = SuperFarmerPKEK::max_stado)
}
\arguments{
\item{kostka1.wynik}{nazwa zwierzecia wylosowanego na pierwszej kostce}

\item{kostka2.wynik}{nazwa zwierzecia wylosowanego na drugiej kostce}

\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{max_stado}{maksymalna liczba zwierzat dostepna w grze, zapisana jako nazwany wektor}
}
\value{
funkcja zwraca StadoGracza po rzucie kostkami
}
\description{
Funkcja reproduction() oblicza stan StadaGracza po rzucie kostkami. Uwzglednia sytuacje, ktorej na kostkach pojawia sie lis lub wilk.
}

