% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3_sprzedaj_kilka_tanszych.R
\name{sprzedaj_kilka_tanszych}
\alias{sprzedaj_kilka_tanszych}
\title{Sprzedawanie kilku tanszych zwierzat i kupowanie drozszego.}
\usage{
sprzedaj_kilka_tanszych(StadoGracza, zwierze, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{zwierze}{zwierze, ktore chcemy kupic}

\item{ceny_w_krolikach}{wektor opisujacy wartosć zwierzat wyrażona w krolikach}
}
\value{
funkcja zwraca StadoGracza po sprzedaniu tanszych zwierzat
}
\description{
Sprzedawanie kilku tanszych zwierzat i kupowanie drozszego.
}

