% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StadoGracza.R
\docType{data}
\name{StadoGracza}
\alias{StadoGracza}
\title{Zbior zawierajacy stan stada gracza, zapisany jako nazwany wektor.}
\description{
Zbior zawierajacy stan stada gracza, zapisany jako nazwany wektor.
}

