% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy1.R
\name{strategy1}
\alias{strategy1}
\title{Strategia 1, wymiana wg strategii 1}
\usage{
strategy1(StadoGracza, max_stado = SuperFarmerPKEK::max_stado,
  ceny_w_krolikach = SuperFarmerPKEK::ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{max_stado}{maksymalna liczba zwierzat dostepna w grze, zapisana jako nazwany wektor}

\item{ceny_w_krolikach}{domyslne wartosci zwierzat wyrazone w krolikach, zapisane jako nazwany wektor}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy1() zwraca StadoGracza po wymianie zgodniej ze strategia 1. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

