% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2.R
\name{strategy2}
\alias{strategy2}
\title{Strategia 2, wymiana wg strategii 2}
\usage{
strategy2(StadoGracza, ceny_w_krolikach = SuperFarmerPKEK::ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{domyslne wartosci zwierzat wyrazone w krolikach, zapisane jako nazwany wektor}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy2() zwraca StadoGracza po wymianie zgodniej ze strategia 2. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

