% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2.StadoGracza.R
\name{strategy2.StadoGracza}
\alias{strategy2.StadoGracza}
\title{Strategia 2, wymiana}
\usage{
strategy2.StadoGracza(StadoGracza, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrazona w krolikach}
}
\value{
funkcja zwraca StadoGracza po wymianie zwierzat zgodnie ze strategia 2
}
\description{
Funkcja strategy_2.StadoGracza() dokonuje wymany zwierzat zgodnie ze strategia 2
}

