% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2.wymiana.krolik.R
\name{strategy2.wymiana.krolik}
\alias{strategy2.wymiana.krolik}
\title{Strategia 2, wymiana kroliki na wieksze zwierze}
\usage{
strategy2.wymiana.krolik(StadoGracza, ceny_w_krolikach, zwierze)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartość zwierzat wyrazona w krolikach}

\item{zwierze}{nazwa zwierzecia, ktore chcemy otrzymać}
}
\value{
funkcja zwraca StadoGracza po wymianie krolikow
}
\description{
Funkcja strategy2.wymiana.krolik() wymiania kroliki na dane wieksze zwierze
}

