% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2.zwierze1_na_mniejsze_zwierze2.R
\name{strategy2.zwierze1_na_mniejsze_zwierze2}
\alias{strategy2.zwierze1_na_mniejsze_zwierze2}
\title{Strategia 2, wymiana jednego zwierzecia na drugie}
\usage{
strategy2.zwierze1_na_mniejsze_zwierze2(StadoGracza, ceny_w_krolikach, zwierze1,
  zwierze2)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartośc zwierzat wyrażona w krolikach}

\item{zwierze1}{nazwa zwierzecia, ktore chcemy wymienic}

\item{zwierze2}{nazwa zwierzecia, ktore chcemy otrzymac}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy2.zwierze1_na_mniejsze_zwierze2() wymiania zwierze1 na zwierze2
}

