% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2_bez_konia.R
\name{strategy2_bez_konia}
\alias{strategy2_bez_konia}
\title{Strategia 2, strategia wymiany w stadzie bez konia}
\usage{
strategy2_bez_konia(StadoGracza, ceny_w_krolikach, wartosc_zagrody, ruch)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrażona w krolikach}

\item{wartosc_zagrody}{liczba wyrażajaca wartosc stada w krolikach}

\item{ruch}{wskaźnik, czy zostala zrobiona wymiana w tym kroku gry (przyjmuje wartosci 0 i 1)}
}
\value{
funkcja zwraca StadoGracza po odpowiedniej wymianie
}
\description{
Funkcja strategy2_bez_konia() opisuje strategie wynian w sytuacji, gdy gracz nie posiada żadnego konia
}

