% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2_mniej_niz_127.R
\name{strategy2_mniej_niz_127}
\alias{strategy2_mniej_niz_127}
\title{Strategia 2, strategia wymiany w stadzie o wartości mniejszej niz 127 krolkow}
\usage{
strategy2_mniej_niz_127(StadoGracza, ceny_w_krolikach, ruch)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartość zwierzat wyrazona w krolikach}

\item{ruch}{wskaznik, czy zostala zrobiona wymiana w tym kroku gry (przyjmuje wartości 0 i 1)}
}
\value{
funkcja zwraca StadoGracza po odpowiedniej wymianie
}
\description{
Funkcja strategy2_mniej_niz_127() opisuje strategie wynian w sytuacji, gdy gracz posiada zwierzeta odpowiadajace mniej niz 127 krolikom
}

