% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2_na_wieksze_zwierze.R
\name{strategy2_na_wieksze_zwierze}
\alias{strategy2_na_wieksze_zwierze}
\title{Strategia 2, wymiana na wieksze zwierze}
\usage{
strategy2_na_wieksze_zwierze(StadoGracza, ceny_w_krolikach, zwierze)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartość zwierzat wyrazona w krolikach}

\item{zwierze}{nazwa zwierzecia, ktore chcemy otrzymać}
}
\value{
funkcja zwraca StadoGracza po zamianie na wieksze zwierze
}
\description{
Funkcja strategy_2_na_wieksze_zwierze() wymiania zwierze 'zwierze'
}

