% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy2_wiecej_niz_jeden_kon.R
\name{strategy2_wiecej_niz_jeden_kon}
\alias{strategy2_wiecej_niz_jeden_kon}
\title{Strategia 2, strategia wymiany w stadzie, z conajmniej 2 konmi}
\usage{
strategy2_wiecej_niz_jeden_kon(StadoGracza, ceny_w_krolikach, ruch)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrażona w krolikach}

\item{ruch}{wskaznik, czy została zrobiona wymiana w tym kroku gry (przyjmuje wartości 0 i 1)}
}
\value{
funkcja zwraca StadoGracza po odpowiedniej wymianie
}
\description{
Funkcja strategy2_wiecej_niz_jeden_kon() opisuje strategie wynian w sytuacji, gdy gracz posiada wiecej niż jednego konia. 
Jest to funkcja konczaca gre
}

