% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.2.R
\name{strategy3.2}
\alias{strategy3.2}
\title{Strategia 3, wymiana wg strategii 3 dla drugiego przedzialu}
\usage{
strategy3.2(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_['krolik']ach i max_stado.
}

