% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.3.R
\name{strategy3.3}
\alias{strategy3.3}
\title{Strategia 3, wymiana wg strategii 3 dla trzeciego przedzialu}
\usage{
strategy3.3(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

