% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.5.R
\name{strategy3.5}
\alias{strategy3.5}
\title{Strategia 3, wymiana wg strategii 3 dla piątego przedzialu}
\usage{
strategy3.5(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

