% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.6.R
\name{strategy3.6}
\alias{strategy3.6}
\title{Strategia 3, wymiana wg strategii 3 dla szostego przedzialu}
\usage{
strategy3.6(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_1ach i max_stado.
}

