% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.7.R
\name{strategy3.7}
\alias{strategy3.7}
\title{Strategia 3, wymiana wg strategii 3 dla siodmego przedzialu}
\usage{
strategy3.7(StadoGracza, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartosć zwierzat wyrażona w krolikach}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

