% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.R
\name{strategy3}
\alias{strategy3}
\title{Strategia 3, wymiana wg strategii 3}
\usage{
strategy3(StadoGracza, ceny_w_krolikach = SuperFarmerPKEK::ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{domyslne wartosci zwierzat wyrazone w krolikach, zapisane jako nazwany wektor}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

