% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3.StadoGracza.R
\name{strategy3.StadoGracza}
\alias{strategy3.StadoGracza}
\title{Strategia 3}
\usage{
strategy3.StadoGracza(StadoGracza, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrazona w krolikach}
}
\value{
funkcja zwraca StadoGracza po wymianie zwierzat zgodnie ze strategia 3
}
\description{
Funkcja strategy_3.StadoGracza() dokonuje wymiany zwierzat zgodnie ze strategia 3
}

