% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3_sprzedaj_drugiego_konia.R
\name{strategy3_sprzedaj_drugiego_konia}
\alias{strategy3_sprzedaj_drugiego_konia}
\title{Strategia 3, wymiana drugiego konia}
\usage{
strategy3_sprzedaj_drugiego_konia(StadoGracza, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{ceny_w_krolikach}{wektor opisujacy wartoĹ›Ä‡ zwierzat wyrazona w krolikach}
}
\value{
funkcja zwraca StadoGracza po zamianie na wieksze zwierze
}
\description{
Funkcja strategy3_sprzedaj_drugiego_konia() wymiania drugiego konia na tansze zwierzeta
}

