% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3_wartosc_stada.R
\name{strategy3_wartosc_stada}
\alias{strategy3_wartosc_stada}
\title{Strategia 3, wartosc stada gracza}
\usage{
strategy3_wartosc_stada(StadoGracza)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}
}
\value{
funkcja zwraca wartosc stada gracza
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

