% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_1.StadoGracza.R
\name{strategy_1.StadoGracza}
\alias{strategy_1.StadoGracza}
\title{Strategia 1, wymiana}
\usage{
strategy_1.StadoGracza(StadoGracza, max_stado, ceny_w_krolikach)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{max_stado}{wektor mowiacy ile zwierzat jest w puli}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrażona w krolikach}
}
\value{
funkcja zwraca StadoGracza po wymianie zwierzat zgodnie ze strategia 1
}
\description{
Funkcja strategy_1.StadoGracza() dokonuje wymany zwierzat zgodnie ze strategia 1
}

