% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_1_na_mniejsze_zwierze.R
\name{strategy_1_na_mniejsze_zwierze}
\alias{strategy_1_na_mniejsze_zwierze}
\title{Strategia 1, wymiana na mniejsze zwierze}
\usage{
strategy_1_na_mniejsze_zwierze(StadoGracza, max_stado, ceny_w_krolikach, i,
  nowa_kolejnosc)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{max_stado}{wektor mowiacy ile zwierzat jest w puli}

\item{ceny_w_krolikach}{wektor opisujacy wartosc zwierzat wyrażona w krolikach}

\item{i}{wspołrzedna, ktora kresla położenie zwierzecia, ktore chcemy wymienic}

\item{nowa_kolejnosc}{kolejnosc uproszczajaca fukcje}
}
\value{
funkcja zwraca StadoGracza po zamianie na mniejsze zwierze
}
\description{
Funkcja strategy_1_na_mniejsze_zwierze() wymiania i-te zwierze
}

