% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_1_na_wieksze_zwierze.R
\name{strategy_1_na_wieksze_zwierze}
\alias{strategy_1_na_wieksze_zwierze}
\title{Strategia 1, wymiana na wieksze zwierze}
\usage{
strategy_1_na_wieksze_zwierze(StadoGracza, max_stado, ceny_w_krolikach, i,
  nowa_kolejnosc)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{max_stado}{wektor mowiacy ile zwierzat jest w puli}

\item{ceny_w_krolikach}{wektor opisujacy wartosć zwierzat wyrażona w krolikach}

\item{i}{wspolrzedna, ktora kresla polożenie zwierzecia, ktore chcemy otrzymać}

\item{nowa_kolejnosc}{kolejnosć uproszczajaca fukcje}
}
\value{
funkcja zwraca StadoGracza po zamianie na wieksze zwierze
}
\description{
Funkcja strategy_1_na_wieksze_zwierze wymiania i-te zwierze
}

