% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizytowkaMEWA.R
\name{wizytowkaMEWA}
\alias{wizytowkaMEWA}
\title{Wizytowka MEWA}
\usage{
wizytowkaMEWA(strategia, name = as.vector(deparse(substitute(strategia)))[1],
  a1 = SuperFarmerPKEK::a1, a33 = SuperFarmerPKEK::a33)
}
\arguments{
\item{strategia}{Strategia, dla ktorej ma zostac wygenerowana wizytowka.}

\item{name}{Nazwa strategii (domyslnie jest to nazwa zmiennej, w ktorej przekazana jest zmienna).}

\item{a1}{wektor czasow dla strategii SuperFarmerDA::strategia_maxrabbit}

\item{a33}{wektor czasow dla strategii SuperFarmerRCNK::strategia_yolo}
}
\description{
Wizytowka danej strategii. Funkcja automatycznie generuje wykresy gestosci i dystrybuant 
oraz tabelki zawierajace podstawowe statystyki. Kazdy z powyzszych elementow wykorzystuje zarowno nasza 
strategie jak i strategie najlepsza (/code{SuperFarmerDA::strategia_maxrabbit}) oraz najgorsza 
(/code{SuperFarmerRCNK::strategia_yolo}).
}
\author{
Ewelina Karbowiak, Marek Wawreniuk
}

