% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy3_wymiana.R
\name{wymiana}
\alias{wymiana}
\title{Wymiana}
\usage{
wymiana(StadoGracza, co_sprzedaj, co_kup, ile_sprzedaj, ile_kup)
}
\arguments{
\item{StadoGracza}{wektor zawieracjacy liczby poszczegolnych zwierzat w stadzie gracza}

\item{co_sprzedaj}{ktore zwierze chcemy sprzedac}

\item{co_kup}{ktore zwierze chcemy kupic}

\item{ile_sprzedaj}{liczba zwierzat, ktore chcemy sprzedac}

\item{ile_kup}{liczba zwierzat, ktore chcemy kupic}
}
\value{
funkcja zwraca StadoGracza po wymianie
}
\description{
Funkcja strategy3() zwraca StadoGracza po wymianie zgodniej ze strategia 3. Zawiera zdefiniowane wektory ceny_w_krolikach i max_stado.
}

